+++ integrationtest/integration_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"os/exec"
 	"strings"
@@ -66,6 +67,23 @@ func TestMount(t *testing.T) {
 
 	_, umount := mountHelper(t, client)
 	defer umount()
+}
+
+func TestCheckViability(t *testing.T) {
+	client, err := api.NewClient(nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	path, umount := mountHelper(t, client)
+	defer umount()
+	client, delNamespace := namespaceHelper(t, client)
+	defer delNamespace()
+
+	// check
+	resp, err := client.Logical().ReadRaw(path + "/check")
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusNoContent, resp.StatusCode)
 }
 
 func TestConfig(t *testing.T) {